fix: sendToWorld Packet

+++ GameServer/Data/Repositories/AccountRepository.cs
@@ -36,14 +36,14 @@ public static class AccountRepository {
 
     private static AccountEntitie MapAccount(MySqlDataReader reader) {
         return new AccountEntitie {
-            Id = reader.GetUInt32("id"),
+            Id = reader.GetUInt16("id"),
             Username = reader.GetString("username"),
             PasswordHash = reader.GetString("passwordHash"),
             Token = reader.GetString("token"),
             TokenCreationTime = reader.GetDateTime("tokenCreationTime"),
-            AccountStatus = reader.GetUInt16("accountStatus"),
-            BanDays = reader.GetUInt16("banDays"),
-            Nation = reader.GetUInt16("nation"),
+            AccountStatus = reader.GetByte("accountStatus"),
+            BanDays = reader.GetByte("banDays"),
+            Nation = reader.GetByte("nation"),
             AccountType = (AccountType)reader.GetInt32("accountType"),
             StorageGold = reader.IsDBNull("storageGold") ? 0 : reader.GetUInt32("storageGold"),
             Cash = reader.IsDBNull("cash") ? 0 : reader.GetUInt32("cash"),
